Move vms_rms.h to internal
