+++ crypto/util/LPdir_vms.c
@@ -51,7 +51,7 @@
 #ifndef LPDIR_H
 # include "LPdir.h"
 #endif
-#include "vms_rms.h"
+#include "internal/vms_rms.h"
 
 /* Some compiler options hide EVMSERR. */
 #ifndef EVMSERR
